requirements.txt
streamlit
tensorflow
ultralytics
pillow
opencv-python-headless
